refactor schema
