+++ src/database/server/schemas/lobechat/_helpers.ts
@@ -1,0 +1,29 @@
+/* eslint-disable sort-keys-fix/sort-keys-fix  */
+import { LobeChatPluginManifest } from '@lobehub/chat-plugin-sdk';
+import { relations } from 'drizzle-orm';
+import {
+  boolean,
+  index,
+  integer,
+  jsonb,
+  pgTable,
+  primaryKey,
+  serial,
+  text,
+  timestamp,
+  unique,
+  uniqueIndex,
+  varchar,
+} from 'drizzle-orm/pg-core';
+import { createInsertSchema, createSelectSchema } from 'drizzle-zod';
+
+import { DEFAULT_PREFERENCE } from '@/const/user';
+import { LobeAgentChatConfig, LobeAgentTTSConfig } from '@/types/agent';
+import { CustomPluginParams } from '@/types/tool/plugin';
+
+import { idGenerator, randomSlug } from '../../utils/idGenerator';
+
+export const timestamptz = (name: string) => timestamp(name, { withTimezone: true });
+
+export const createdAt = () => timestamptz('created_at').notNull().defaultNow();
+export const updatedAt = () => timestamptz('updated_at').notNull().defaultNow();
